Reduce the friction in getting these engines talking to each other.
There are a few things this commit covers:

= Properly allowing Alchemy content in Solidus views

The existing instructions recommended the following be added to allow
alchemy content to be available in Solidus views:

    Spree::BaseController.class_eval do
      include Alchemy::ControllerActions
    end
    Spree::UserSessionsController.class_eval do
      include Alchemy::ControllerActions
    end

This recommendation lead to Issue 4 on Github. This causes `current_ability`
to return Alchemy's ability even when in a Solidus controller which was not
compatible with Solidus (even though the ability was merged into Alchemy).

Taking a queue from a hack in Narrator[1] (a Spree/Alchemy integration) I
was able to include the Alchemy controller helpers into Spree except the
`current_abiltiy` method. Should it be needed it is still aliased as
`current_alchemy_ability`. Since this is slightly complicated I wrapped
this up into the file `alchemy_in_solidus` allowing for an easy include in
the initializer to do it right. This allows the integration to still be
optional but easier to do without error and closes issues like #4.

Also the require automatically detects if solidus-auth-devise is being used
and only does the include conditionally removing one more thing for the
integrator to think about.

= Simplifying integration with the Solidus Layout

I liked how the block of code was replaced with just a simple require so
I did the same thing for allowing Alchemy to render in Spree's layout. The
README recommended the following code:

    Alchemy::BaseHelper.send :include, Spree::BaseHelper
    Alchemy::BaseHelper.send :include, Spree::StoreHelper
    Alchemy::BaseController.send :include, Spree::Core::ControllerHelpers::Common
    Alchemy::BaseController.send :include, Spree::Core::ControllerHelpers::Store

This is now replace with just:

    require 'alchemy/solidus/use_solidus_layout'

Again still optional but now much less code.

= Automatic User Decoration

There were two different decorators for the different user models that needed to
be required. The system now automatically detects the selected user model and
requires the decorator for the developer. Again one less thing to think about.

In addition the previous README had the require directly in the initializer. It is
not in the `to_prepare` callback making it work with reloading better.

= Updated README

Finally I updated the readme to provide the now more simplfied integration instructions.
In addition I included some information about how to setup the routing to provide
the integrator easy copy/paste code.

1. https://github.com/boomerdigital/spree_narrator/blob/master/config/initializers/alchemy.rb#L30